python example: automatic passwordless ssh
passwordless ssh is useful for system administrators.

While when there are huge amount of machines you want to manager, this
process will take a lot of time to enable this. This script is used to
make the process automatically.

This script will be provided with a machine list file which contains the
ip address and root password.