resource.c: add more comment on __request_region()
some comment on __request_region and one example on this function is
added.

especially those comments and example are used to compare the difference
between __request_region() and __request_resource().

the __request_resource() will just request resource on the first level
of parent while __request_region() will go deep in the tree.