kernel_init: how init process is invoked, in initramfs or in root?
In kernel_init(), sys_access() is called to verify whether /init or other init
program is in the initramfs.

If this program can't be located, kernel_init() will call prepare_namespace()
to mount the hard disk to be the root fs.

If this program is there, then kernel will let it do all the work. In a normal
PC or hmm... server case, this /init will do some initialization and then
mount the rootfs on harddisk and exec the init process on harddisk. This is
the process 1 after we login the system and can be seen with ps command.