kernel makefile: one test on the kernel build system
My intention is to create a directory and put it under the root directory of
the kernel source tree.

This directory will be built during the process of kernel building.

Add this directory in the root directory, and make following change in the
Makefile in the root directory of kernel.

 ifeq ($(KBUILD_EXTMOD),)
-core-y         += kernel/ mm/ fs/ ipc/ security/ crypto/ block/
+core-y         += kernel/ mm/ fs/ ipc/ security/ crypto/ block/ test/