python example: thread in python
In python, thread is implemented by threading library. Normally, we create a
class which is inherited from the threading and implement the run function.

When we want to create a thread, we instanciate an object and call the start
function.

In this example, we can see the thread is running at the same time when the
main function is running.