kernel code reading: raise a softirq
When raise a softirq, kernel just set the related pending bit on this cpu.
So it is quick.

And combined with the operation of do_softirq(), this show softirq will be
handled on the exact cpu who raised it. This means if one cpu is busy with
interrupt, it will be busy with softirq too. One solution is to use interrupt
affinity to distribute interrupt to several cpus.