kernel code reading: those relationships super_block/dentry/inode
dentry creates a tree hierarchy, with each dentry connects to an inode.

One super_block has a pointer to the root dentry, which is the root of the
dentry tree. By doing so, it holds the whold picture of this partition.

Meanwhile, super_block maintains a list of all those inodes in this partition.

Another questions is, in the dentry tree, each node could represent a
file/directory/symlink?