kernel code reading: add a flag to case list_for_each_entry_xxx()
If there is no flag, each time we invoke these function will re-init the whole
list. While actually, we just want to dump the list content.

This one fix this issue by adding a flag. When it is 1, it will re-init the
list. When it is 0, it just dump the list.