kernel code reading: blk's elevator
Elevator is a kind of mechanism in block layer to optimize the IO.

Each request_queue will be assigned an elevator. From the request_queue point
of view, elevator is a helper to order its requests. When the someone ask the
request_queue for a request to process, the request_queue will ask the
elevator to give him one.

This is my point of view.