kernel/kobj: use ktype->sysfs_ops correctly
In previous example, the sysfs_ops usage is not correct. Actually, it is
called on each entry of the default_attrs. Each entry in default_attrs will
create a file in sysfs directory. And when the sysfs_ops is called, the
"attribute" will be passed.

Here is the first example which is borrowed from fs/ext4/sysfs.c. In this
example, it use an attr_id to distinguish which file it is.