kernel makefile: FORCE target
A rule with no prerequisites or recipe, and the target of the rule is a
nonexistent file, then make imagines this target to have been updated whenever
its rule is run. This implies that all targets depending on this one will
always have their recipe run.

So we can add this to the rule we want to force it to run every time.