qemu: correct way to use ipvtap in guest
By default, ens3 would has capability to carry ARP packets.

Since its backed is a L3 ipvtap interface, the ARP packets would be dropped.
This is the reason for the trick to add a route to host and set static ARP
address in Version 1.

Now the correct way is to disable ARP for ens in guest.

After doing so, the route table in guest is must clear.