userspace/c: tun interface to fake ICMP reply
Steps:

$ ./tun -n tun0 -t tun
$ ip link set tun0 up
$ ip addr add  192.168.3.11/24 dev tun0
$ ping -c 1 192.168.3.12

The ip 192.168.3.12 doesn't exist, while still get ICMP reply.
This is because the tun.c fake the ICMP reply and reply.

!!!Note:

This example just work for tun device, since current modification is based on
IP packet format which is the tun device read format. In case this is a tap
device, another adjustment is necessary.