userspace/bazel: load your private bazel rule
In commit (83f4bb5 userspace/bazel: use container_push() to push to private
registry), we take an approach to modify bazel rule in local cache. While this
is not convenient.

Actually, we can upload the compressed code to http server and then use it any
where.

Two ways to use it:

  1. Upload to github as show in this patch
     "https://github.com/RichardWeiYang/rules_docker/releases/download/v0.14.4-1/rules_docker-v0.14.4-1.tar.gz",
  2. Setup you own http server
     "http://localhost:8000/rules_docker-0.14.4.tar.gz",

Note:

    The suffix is important, .tar.gz. And the compress command is

    $ tar czf rules_docker-0.14.4.tar.gz rules_docker

Note2:

    The stript_prefix should be changed to "rules_docker".
    Not sure how to adjust is in .tar.gz

Note3:

    When using "http", we should add the sha256 value.

    $ sha256sum rules_docker-0.14.4.tar.gz