userspace/bash: get remaining argument
I just found a limitation of getopts.

The final parameter must be put at the end.
Or what I should say is getopts stops when it encounter the first
non-option.

Compare the following two command line:

  ./getopts.sh -p path 88
  ./getopts.sh 88 -p path

For the first one, we will parse option "-p" correctly, while for the
second one, it won't.